Bulk Upload Phase 1 Release Fixes

+++ pipeline/alembic/versions/37a01265c977_add_gender_imported_industrial_athlete.py
@@ -7,7 +7,7 @@ Create Date: 2020-08-11 19:42:37.427735
 """
 from alembic import op
 import sqlalchemy as sa
-
+from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
 revision = "37a01265c977"
@@ -24,4 +24,9 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_column("imported_industrial_athlete", "gender")
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+    
+    if 'imported_industrial_athlete' in tables:
+        op.drop_column("imported_industrial_athlete", "gender")
